[面試表單] 將題目抽出成可共用的creators (#1199)
title:	[面試表單] 將題目抽出成可共用的creators
state:	OPEN
author:	peteranny
labels:	
assignees:	
reviewers:	barry800414 (Approved)
projects:	
milestone:	
number:	1199
url:	https://github.com/goodjoblife/GoodJobShare/pull/1199
additions:	375
deletions:	323
auto-merge:	disabled
--
#1195  <!-- 當 PR merge，github 會自動幫你關 issue -->

## 這個 PR 是？ <!-- 必填 -->

面試的題目很多是可與其他表單共用的
這個PR把這些題目先抽出放在共用處
其他表單後續可以沿用

## 我應該如何手動測試？ <!-- 必填 -->

- [ ] https://www.goodjob.life/share 面試表單應不受影響
